project/requirements.txt
Flask==0.12.2
google-cloud==0.26.1
google-api-python-client==1.6.2
mailchimp3==2.0.15
requests[security]==2.18.1
urllib3[secure]==1.22
